Reconciled StorageManager File/Block Variables 	- make _Everything_ Uint64, lest we introduce any ugly math issues 	in computing and comparing things... 		-> we're seeing issues on a few beamlines with enforcement 		of the "Max Blocks Allowed" setting... ;-Q 	- m_block_size: "uint32_t" -> "uint64_t" 		-> in StorageManager class 		-> in PoolsizePV class 		-> in PercentPV class 	- m_percent: int -> uint32_t 		-> in StorageManager class 		-> local "percent" in PercentPV class changed() method... 	- change "size" parameter to StorageManager::addBaseStorage() 		-> "off_t" -> "uint64_t" 			=> likewise with local "blocks" 	- in StorageManager::addPacket() 		-> locals "size" and "blocks": "off_t" -> "uint64_t" 	- in StorageManager::savePacket() 		-> locals "size" and "blocks": "off_t" -> "uint64_t" 	- cast "percent" values when calculating "Max Size" (in Uint64)... 		-> in PercentPV class changed() method... 		-> in StorageManager::config() 	- cast stat() struct stat "st_blksize" field to Uint64 to be sure... 		-> in StorageManager::config() 	- cast fsstats struct "f_blocks" field to Uint64 to be sure... 		-> in StorageManager::config() 		-> in StorageManager::set_max_blocks_allowed()

+++ sms/StorageManager.h
@@ -135,9 +135,9 @@ private:
 	static int m_base_fd;
 
 	static std::string m_poolsize;
-	static int m_percent;
+	static uint32_t m_percent;
 
-	static uint32_t m_block_size;
+	static uint64_t m_block_size;
 	static uint64_t m_blocks_used;
 	static uint64_t m_max_blocks_allowed;
 
@@ -198,7 +198,7 @@ private:
 				bool last);
 	static void populateDailyCache(void);
 
-	static void addBaseStorage(off_t size);
+	static void addBaseStorage(uint64_t size);
 	static void startContainer(uint32_t run = 0,
 				std::string propId = std::string("UNKNOWN"));
 	static void endCurrentContainer(void);
